pyproject.toml
[project]
name = "meal-prep-agent"
version = "0.1.0"
description = "LangGraph-based meal prep agent with Tesco API integration"
authors = [
    {name = "Matthew Sharp"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "langgraph (>=0.5.0,<0.6.0)",
    "langchain-core (>=0.3.66,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "grandalf (>=0.8,<0.9)",
    "requests (>=2.32.4,<3.0.0)",
    "playwright (>=1.53.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
